% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ann.partition.model.1}
\alias{ann.partition.model.1}
\title{Data: ANN model for use in component partitioning}
\format{An object of class "nn" built using the neuralnet package}
\source{
See PMID:(TBD) for details of ann construction.
}
\usage{
ann.partition.model.1
}
\description{
An artificial neural network (ANN) trained using simulated data to find the
optimal thresholds for partitioning components into discrete sets.
}
\keyword{datasets}
