% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{ann.partition}
\alias{ann.partition}
\title{Wrapper function for ANN partitioning}
\usage{
ann.partition(x, model = ann.partition.model.1)
}
\arguments{
\item{x}{A numeric matrix}

\item{model}{ANN model to use for partitioning}
}
\value{
A matrix comprising values of 1, 0, and -1 to indicate column-wise
  partitioning of the input matrix into positive, neither, and negative sets,
  respectively.
}
\description{
Wrapper function for \code{\link{partition}} for partitioning using an
artificial neural network
}
\details{
The ("\code{model}") parameter specifies the ANN model to use for
  calculating thresholds.  The default values is
  \code{ann.partition.model.1}, which is supplied with this package.
}
\examples{
x = matrix(rnorm(9), 3, 3)
ann.partition(x)
}
