% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkModules.R
\name{checkModules}
\alias{checkModules}
\title{Evaluate a set of modules using gene annotations}
\usage{
checkModules(x, a, p = 0.05, return.p.matrix = FALSE)
}
\arguments{
\item{x}{a binary matrix indicating module membership (genes in rows, modules in columns)}

\item{a}{a binary gene annotation matrix (genes in rows, annotations in columns)}

\item{p}{FDR-adjusted p-value below which to consider an annotation significant}

\item{return.p.matrix}{logical value indicating whether the full p-value matrix should
be returned in the output}
}
\value{
A list with the following elements:
\describe{
 \item{num.annotations}{Number of annotations tested (number of columns in a)}
 \item{num.modules}{Number of modules tested (number of columns in x)}
 \item{anns.signif}{Number of annotations that were significant in at least
 one module}
 \item{mods.signif}{Number of modules that were significant for at least
 one annotation}
 \item{p.matrix}{The module-by-annotation p-value matrix (if return.p.matrix == TRUE)}
}
}
\description{
Evaluates a set of gene modules (e.g., the output from
\code{\link{predictModules}}) by checking them for enrichment of gene
annotations.
}
\details{
This function evaluates a set of predicted gene modules by testing whether
each module is enriched for any gene annotations.  Enrichment p-values are calculated using
hypergeometric statistics (see \code{\link[stats]{phyper}}).  If x is not a binary matrix,
module membership is determined using \code{\link[DEXICA]{partitionComponents}} with default
paramaters.  The FDR-correction (see \code{\link[stats]{p.adjust}}) is applied to all
p-values prior to determining signficance.
}
\examples{
x = matrix(rnorm(100), 10, 10)
a = matrix(sample(c(0,1), size = 100, replace = TRUE), 10, 10)
m = predictModules(x)$S
rownames(a) = rownames(m) = as.character(1:10)
m.p = partition(m, t = 2)
m.p2 = splitMatrix(m.p)
checkModules(m.p2, a)

}
