% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexBatch.R
\name{countJobs}
\alias{countJobs}
\title{Count the number of jobs defined in a DexBatch object}
\usage{
countJobs(object)
}
\arguments{
\item{object}{Object to get job count for.}
}
\description{
This function returns the expected number of individual jobs defined in a
DexBatch object.  The formula it uses is \code{c * p}, where c is the number of
compendia and p is the number of different parameter sets to be tested.
}
\examples{
x = matrix(rnorm(10 * 10), 10, 10)
y = matrix(rnorm(10 * 10), 10, 10)
a1 = matrix(sample(c(0,1), 100, replace = TRUE))
db = dexBatch(list(x,y), list(a1))
countJobs(db)

}
