% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexBatch.R
\docType{class}
\name{DexBatch-class}
\alias{DexBatch-class}
\alias{dexBatch}
\alias{countJobs,DexBatch-method}
\alias{runJob,DexBatch-method}
\title{An S4 class for testing multiple compendia and/or parameters}
\usage{
## Constructor:
dexBatch(compen, annmats, params = NULL, output = NULL)

\S4method{countJobs}{DexBatch}(object)

\S4method{runJob}{DexBatch}(object, j)
}
\description{
The DexBatch class helps users test the effect of different
parameter options and/or input compendia on gene module prediction.
}
\details{
A DexBatch object defines a set of jobs.  Each job first predicts a set of
gene modules, then evaluates their quality.  Jobs within a DexBatch differ from
each other in some way; the purpose of a DexBatch is to test the effect of
these differences on the quality of the resulting modules.

Individual jobs in a DexBatch are executed by calling \code{\link[DEXICA]{runJob}}.
While each job could be run sequentially on a single machine, it is usually
more practical when testing a large number of parameters to execude the jobs in
parallel (see Use Cases in the DEXICA vignette)

If params is NULL, a single parameter set will be supplied using default values for
the \code{\link[DEXICA]{parameterSet}} constructor.  If output is NULL, results will
be directed to \code{stdout()}.
}
\section{Methods (by generic)}{
\itemize{
\item \code{countJobs}: Count the number of jobs defined in a DexBatch object (see
\code{\link[DEXICA]{countJobs}})

\item \code{runJob}: Run a single job defined in a DexBatch object (see
\code{\link[DEXICA]{runJob}})
}}

\section{Slots}{

\describe{
\item{\code{compen}}{A matrix or list of matrices containing the compendia to be tested}

\item{\code{annmats}}{A matrix or list of matrices containing the annotation matrices to be used
in module evaltation (see \code{\link[DEXICA]{checkModules}})}

\item{\code{params}}{A ParameterSet object containing desired parameter ranges to be
tested (see \code{\link[DEXICA]{parameterSet}})}

\item{\code{output}}{Character string name of output file}
}}

