% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{fixed.partition}
\alias{fixed.partition}
\title{Wrapper function for fixed threshold partitioning}
\usage{
fixed.partition(x, t = 3)
}
\arguments{
\item{x}{A numeric matrix}

\item{t}{Threshold value for use in partitioning.  Values in x >= t will
become 1 in the return matrix, values in x <= -t will become -1, and values
in x between -t and t will become 0}
}
\value{
A matrix comprising values of 1, 0, and -1 to indicate column-wise
  partitioning of the input matrix into positive, neither, and negative sets,
  respectively.
}
\description{
Wrapper function for \code{\link{partition}} for fixed threshold partitioning
}
\examples{
x = matrix(rnorm(9), 3, 3)
fixed.partition(x)
}
