% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{getAllParamSets}
\alias{getAllParamSets}
\title{Get all parameter sets defined in a ParameterSet object}
\usage{
getAllParamSets(object)
}
\arguments{
\item{object}{Object to get parameter sets from.}
}
\description{
This function returns all parameter sets defined in a
ParameterSet object.
}
\examples{
p = parameterSet(n.comp = c(10,20,30))
getAllParamSets(p)

}
