% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{getOneParamSet}
\alias{getOneParamSet}
\title{Get a single parameter set defined in a ParameterSet object}
\usage{
getOneParamSet(object, j)
}
\arguments{
\item{object}{Object to get parameter set from.}

\item{j}{An integer indicating which parameter set in object to return}
}
\description{
This function returns a single parameter sets defined in a
ParameterSet object.
}
\examples{
p = parameterSet(n.comp = c(10,20,30))
getOneParamSet(p, 1)

}
