% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{getParamSetCount}
\alias{getParamSetCount}
\title{Get the number of parameter sets defined in a ParameterSet object}
\usage{
getParamSetCount(object)
}
\arguments{
\item{object}{Object to get parameter set count for.}
}
\description{
This function returns the number of individual parameter sets defined in a
ParameterSet object.
}
\examples{
p = parameterSet(n.comp = c(10,20,30))
getParamSetCount(p)

}
