% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GPL200.GO.mini}
\alias{GPL200.GO.mini}
\title{Data: A binary matrix associating probesets with Gene Ontology terms}
\format{A binary matrix with probeset IDs in rows and GO IDs in columns.}
\usage{
GPL200.GO.mini
}
\description{
This is a binary probeset annotation matrix containing Affymetrix probeset IDs in rows
and Gene Ontology (GO) IDs in columns. A value of 1 in the matrix indicates that the
gene to which the probeset has been mapped has been annotated with the corresponding
GO term.  This matrix is a subset of the \code{GPL200.GO}
matrix from the DEXDATA.Celegans package and is intended for use in example scripts.
The matrix was constructed by first limiting the rows (probesets) to those found in
\code{\link[DEXICA]{GPL200.50.mini}}, then by selecting 1000 columns (GO IDs) at random
from those that contained at least 5 annoted probesets.
}
\keyword{datasets}
