% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\docType{class}
\name{ParameterSet-class}
\alias{ParameterSet-class}
\alias{parameterSet}
\alias{getParamSetCount,ParameterSet-method}
\alias{getOneParamSet,ParameterSet-method}
\alias{getAllParamSets,ParameterSet-method}
\alias{reseedParamSets,ParameterSet-method}
\title{An S4 class for defining DexBatch parameter ranges}
\usage{
## Constructor:
parameterSet(n.comp = 1, center.cols = c(TRUE), scale.cols = c(FALSE),
center.rows = c(FALSE), scale.rows = c(FALSE), simultaneous = c(FALSE),
alg.typ = "parallel", fun = "logcosh", alpha = 1,
maxit = 200, tol = 1e-04, w.init = NULL, max.attempts = 3, verbose = FALSE)

\S4method{getParamSetCount}{ParameterSet}(object)

\S4method{getOneParamSet}{ParameterSet}(object, j)

\S4method{getAllParamSets}{ParameterSet}(object)

\S4method{reseedParamSets}{ParameterSet}(object, new.seeds = NULL)
}
\description{
A ParameterSet object defines a range of parameter values to
test in a DexBatch (see \code{\link[DEXICA]{DexBatch-class}}).
}
\details{
When creating a DexBatch object, a ParameterSet may
be supplied (in the \code{params} slot) in order to specify a range of parameter
values to test.  Each slot in ParameterSet may accept either a single value or a
range of values.  If multiple slots have multiple values, the ParameterSet will
encompass each possible combination of individual slot values.

Any slot values not specified during object construction are given their default
values (see constructor), e.g., calling \code{parameterSet()}
results in a ParameterSet with a single set of parameter values.  The
\code{\link[DEXICA]{getParamSetCount}} method returns the total number of
parameter sets defined in a ParameterSet.
}
\section{Methods (by generic)}{
\itemize{
\item \code{getParamSetCount}: Get the number of parameter sets defined in a ParameterSet object (see
\code{\link[DEXICA]{getParamSetCount}})

\item \code{getOneParamSet}: Get a single parameter set defined in a ParameterSet object (see
\code{\link[DEXICA]{getOneParamSet}})

\item \code{getAllParamSets}: Get all parameter sets defined in a ParameterSet object (see
\code{\link[DEXICA]{getAllParamSets}})

\item \code{reseedParamSets}: Reset the random seeds in a ParameterSet object (see
\code{\link[DEXICA]{reseedParamSets}})
}}

\section{Slots}{

\describe{
\item{\code{n.comp}}{an integer or vector specifying the number of components to extract}

\item{\code{alg.typ,fun,alpha,maxit,tol,w.init,max.attempts,verbose}}{See \code{\link[DEXICA]{dexFastICA}} documentation}

\item{\code{center.cols,scale.cols,center.rows,scale.rows,simultaneous}}{See \code{\link[DEXICA]{preprocessMatrix}} documentation}
}}

