% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partition.R
\name{partition}
\alias{partition}
\title{Partition a component matrix}
\usage{
partition(x, method = c("fixed", "ann"), ...)
}
\arguments{
\item{x}{A numeric matrix}

\item{method}{Specifies method to use for determining partitioning threshold values}

\item{...}{Optional arguments to partition method}
}
\value{
A matrix comprising values of 1, 0, and -1 to indicate column-wise
  partitioning of the input matrix into positive, neither, and negative sets,
  respectively.
}
\description{
Given a numeric matrix, returns a matrix with the same dimensions, row and
column names but with only 1, 0, and -1 as values.  These values indicate the
partitioning of each column of the matrix into positive, neither, and
negative subsets.
}
\details{
The ("\code{fixed}") method uses a fixed value (specified by an additional
parameter, \code{t}, which has a default value of 2.8) to create  partitions, while
the ("\code{ann}") method uses an artificial neural network model to create custom
partition thresholds for each component.
}
\examples{
x = matrix(rnorm(9), 3, 3)
partition(x, method = "fixed", t = 0.5)
}
