% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParameterSet.R
\name{reseedParamSets}
\alias{reseedParamSets}
\title{Reset the random seeds in a ParameterSet object}
\usage{
reseedParamSets(object, new.seeds = NULL)
}
\arguments{
\item{object}{Object in which to reset random seeds}

\item{new.seeds}{Numeric vector indicating new random seeds}
}
\description{
This function resets the random seeds defined in a ParameterSet
(specifically, in the \code{w.init} slot) to random values if new.seeds is NULL,
or to the values specified in new.seeds otherwise.
}
\details{
If new.seeds is not NULL, it must be a numeric vector with a length
equal to the number of parameter sets defined in \code{object} (this value
can be determined with the \code{\link[DEXICA]{getParamSetCount}} method.)
}
\examples{
p = parameterSet(n.comp = c(10,20,30))
getAllParamSets(p)

}
