% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DexBatch.R
\name{runJob}
\alias{runJob}
\title{Run a single job defined in a DexBatch object}
\usage{
runJob(object, j)
}
\arguments{
\item{object}{DexBatch object that defines job parameters.}

\item{j}{Job number to run}
}
\description{
This function executes a single job defined in a DexBatch object and writes
the results to the output connection defined therein.
}
\examples{
x = matrix(rnorm(10 * 10), 10, 10)
y = matrix(rnorm(10 * 10), 10, 10)
a1 = matrix(sample(c(0,1), 100, replace = TRUE))
db = dexBatch(list(x,y), list(a1))
runJob(db, 1)

}
