% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitMatrix.R
\name{splitMatrix}
\alias{splitMatrix}
\title{Split matrix}
\usage{
splitMatrix(x)
}
\arguments{
\item{x}{A numeric matrix}
}
\value{
A matrix devoid of negative values with twice the number of columns as
the input matrix.
}
\description{
Given a numeric matrix, this function performs a column-wise separation of positive
values from negative values, producing a return matrix with twice the number of columns
as the input matrix.
}
\details{
Conceptually, this function makes two copies of the input matrix (we will call
them A and B).  It replaces all negative values in copy A with 0, and all positive
values in copy B with 0.  It then multiplies copy B by -1, and finally combines (via \code{\link{cbind}})
the two copies together column-wise.  The strings "_pos" and "_neg" are appended to the
original column names to generate the column names for the new matrix.
}
\examples{
x = matrix(rnorm(9), 3, 3)
splitMatrix(x)
}
