% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessMatrix.R
\name{preprocessMatrix}
\alias{preprocessMatrix}
\title{Preprocess a data matrix}
\usage{
preprocessMatrix(x, center.cols = TRUE, scale.cols = FALSE, center.rows = FALSE,
scale.rows = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a numeric matrix(-like object), typically with genes in rows and samples
(e.g., microarrays) in columns}

\item{center.cols}{a logical value indicating whether to adjust columns to have 0 mean}

\item{scale.cols}{a logical value indicating whether to adjust columns to have unit variance}

\item{center.rows}{a logical value indicating whether to adjust rows to have 0 mean}

\item{scale.rows}{a logical value indicating whether to adjust rows to have unit variance}

\item{verbose}{a logical value indicating the desired level of output as the algorithm runs}
}
\value{
A centered and/or scaled matrix.
}
\description{
Centers and/or scales the columns and/or rows of the input matrix.
}
\details{
This function centers and/or scales the columns and/or rows of the input matrix.
Default parameter values correspond to those of the fastICA algorithm.
}
\examples{
x = matrix(rnorm(100), 10, 10)
x = preprocessMatrix(x)

}
\seealso{
\code{\link[base]{scale}}
}
