% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleMatrix.R
\name{scaleMatrix}
\alias{scaleMatrix}
\title{Scale rows and columns of a matrix}
\usage{
scaleMatrix(x, center.rows = TRUE, center.cols = TRUE, scale.rows = TRUE,
scale.cols = TRUE, steps = 1000, bad.row.action = c("resample", "remove"),
bad.col.action = c("resample", "remove"), set.na.to = 0, verbose = FALSE)
}
\arguments{
\item{x}{a numeric matrix(-like object)}

\item{center.rows}{a logical value indicating whether to set row means to 0}

\item{center.cols}{a logical value indicating whether to set column means to 0}

\item{scale.rows}{a logical value indicating whether to set row variances to 1}

\item{scale.cols}{a logical value indicating whether to set column variances to 1}

\item{steps}{an integer}

\item{bad.row.action}{action to take when a row in x has 0 variance}

\item{bad.col.action}{action to take when a column in x has 0 variance}

\item{set.na.to}{numeric value to set NA values in x to}

\item{verbose}{a logical value indicating whether to show status messages during execution}
}
\value{
A scaled and/or centered matrix.
}
\description{
Centers and/or scales the rows and/or columns of a numeric matrix.
}
\details{
This function is a generalization of the generic function \code{\link[base]{scale}}, which
operates on only the columns of a matrix.  This function (optionally) operates on both
the rows and columns simultaneously, providing a resulting matrix with both rows and
columns having zero means and unit variance.

The algorithm converges toward the solution in a stepwise manner - the larger the number
of steps, the more accurate (i.e., the closer to the desired outcome) will be the final
solution.  A step size of 1000 resulted in convergence (deviation from target < 2.2e-16)
in most test cases, but may be time consuming on large matrices.  Lower values for
\code{steps} will execute faster, but with possible loss of accuracy; a \code{step} size
< 100 is not recommended.

If either: a) both \code{center.cols} and \code{scale.cols} are \code{FALSE}, or b) both
\code{center.rows} and \code{scale.rows} are \code{FALSE}, the function defaults to the
generic \code{\link[base]{scale}} function (using transposition to handle the latter case).
}
\examples{
x = matrix(rnorm(100), 10, 10)
range(apply(x, 1, mean)) # row means
range(apply(x, 1, sd)) # row sd
range(apply(x, 2, mean)) # column means
range(apply(x, 2, sd)) # column sds

x.scaled = scaleMatrix(x)
range(apply(x.scaled, 1, mean)) # row means
range(apply(x.scaled, 1, sd)) # row sds
range(apply(x.scaled, 2, mean)) # column means
range(apply(x.scaled, 2, sd)) # column sds

}
